All tests passed. Comments can be a little cleaner.

+++ src/main/java/com/example/repository/AccountRepository.java
@@ -4,13 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.entity.Account;
 
-public interface AccountRepository extends JpaRepository<Account, Long> {
-    /*
-     * public Account insertAccount(Account account);
-     * public Account getAccount(String username, String password);
-     * public Account getAccountById(int account_id);
-     * public boolean doesUsernameExist(String username); //should have been Account
-     */
+public interface AccountRepository extends JpaRepository<Account, Integer> {
     Account findAccountByUsername(String username);
     Account findAccountByUsernameAndPassword(String username, String password);
 }
